+++ app/controllers/supporting_documents_controller.rb
@@ -1,0 +1,62 @@
+# frozen_string_literal: true
+
+# Allows users to upload supporting documents to a data request.
+class SupportingDocumentsController < ApplicationController
+  before_action :authenticate_user!
+  before_action :find_data_request_or_redirect
+  before_action :find_supporting_document_or_redirect, only: [:show, :destroy]
+
+  # GET /data/requests/:data_request_id/supporting-documents
+  def index
+    @supporting_documents = @data_request.supporting_documents # TODO: Paginate or remove.
+  end
+
+  # GET /data/requests/:data_request_id/supporting-documents/1
+  def show
+    send_file @supporting_document.document.path, disposition: "inline"
+  end
+
+  # GET /data/requests/:data_request_id/supporting-documents/new
+  def new
+    @supporting_document = @data_request.supporting_documents.new
+  end
+
+  # POST /data/requests/:data_request_id/supporting-documents
+  def create
+    @supporting_document = @data_request.supporting_documents.new(supporting_document_params)
+    if @supporting_document.save
+      redirect_to data_requests_addons_path(@data_request), notice: "Supporting document was successfully created."
+    else
+      render :new
+    end
+  end
+
+  # POST /data/requests/:data_request_id/supporting-documents/upload.js
+  def create_multiple
+    params[:documents].each do |document|
+      @data_request.supporting_documents.create(document: document)
+    end
+    render :index
+  end
+
+  # DELETE /data/requests/:data_request_id/supporting-documents/1.js
+  def destroy
+    @supporting_document.destroy
+  end
+
+  private
+
+  def find_data_request_or_redirect
+    @data_request = current_user.data_requests.submittable.find_by(id: params[:data_request_id])
+    empty_response_or_root_path unless @data_request
+  end
+
+  def find_supporting_document_or_redirect
+    @supporting_document = @data_request.supporting_documents.find_by(id: params[:id])
+    empty_response_or_root_path unless @supporting_document
+  end
+
+  def supporting_document_params
+    params.require(:supporting_document).permit(:document)
+  end
+end