+++ app/models/supporting_document.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+# Stores PDFs and images attached to data requests.
+class SupportingDocument < ApplicationRecord
+  # Uploaders
+  mount_uploader :document, SupportingDocumentUploader
+
+  # Validations
+  validates :document, presence: true
+
+  # Relationships
+  belongs_to :data_request, class_name: "Agreement"
+end