+++ test/system/supporting_documents_test.rb
@@ -1,0 +1,9 @@
+require "application_system_test_case"
+
+class SupportingDocumentsTest < ApplicationSystemTestCase
+  # test "visiting the index" do
+  #   visit supporting_documents_url
+  #
+  #   assert_selector "h1", text: "SupportingDocument"
+  # end
+end